requirements.txt
Django==6.0
djangorestframework==3.16.1
google-genai==1.56.0
python-dotenv==1.2.1
drf_spectacular==0.29.0